fix: remove duplicate widget folder
